Create RawSelectInput
- Separate searching mechanism from SearchInput
