+++ v2/Input/SelectInput/index.tsx
@@ -1,33 +1,25 @@
 import React, {
-    useCallback,
-    useEffect,
     useMemo,
-    useRef,
     useState,
 } from 'react';
 import {
-    _cs,
+    isFalsyString,
     caseInsensitiveSubmatch,
     compareStringSearch,
-    isFalsyString,
-    isTruthyString,
 } from '@togglecorp/fujs';
 
-import {
-    calcFloatPositionInMainWindow,
-    defaultLimit,
-    defaultOffset,
-} from '../../../utils/bounds';
 import { OptionKey } from '../../types';
 
-import Button from '../../Action/Button';
-import useKeyboard from '../../General/useKeyboard';
-import RawInput from '../RawInput';
-import HintAndError from '../HintAndError';
-import Label from '../Label';
+import RawSelectInput, { RawSelectInputProps } from '../RawSelectInput';
 
-import Options from './Options';
-import styles from './styles.scss';
+interface DefaultItem {
+    key: string;
+    label: string;
+}
+
+interface State<K> {
+    searchValue?: string;
+}
 
 /*
 # Feature
@@ -36,6 +28,7 @@ import styles from './styles.scss';
 - Clicking on dropdown button will toggle popup visibility
 - Can hide/un-hide dropdown button using prop showDropdownArrowButton
 - Support string or number as key
+- Create RawSelectInput to pull data from server
 
 # Breaking Change
 - Remove prop focusedKey
@@ -46,379 +39,75 @@ import styles from './styles.scss';
 - Show values that are invalid (tally with current options)
 */
 
-interface DefaultItem {
-    key: string;
-    label: string;
-}
-
-interface State<K> {
-    inputInFocus?: boolean;
-    focusedKey?: K;
-    showOptionsPopup: boolean;
-    searchValue?: string;
-}
-
-interface Props<T, K extends OptionKey> {
-    autoFocus?: boolean;
-    className?: string;
-    disabled: boolean;
-    error?: string;
-    hint?: string;
-    keySelector: (datum: T) => K;
-    label?: string;
-    labelSelector: (datum: T) => string | number;
-    onChange: (key: K | undefined) => void;
-    onOptionClick: (key: K) => void;
-    onOptionFocus: (key: K) => void;
-    optionLabelSelector?: (datum: T) => React.ReactNode;
-    options: T[];
-    optionsClassName?: string;
-    placeholder?: string;
-    readOnly: boolean;
-    renderEmpty: React.ComponentType<unknown>;
-    showClearButton: boolean;
-    showDropdownArrowButton: boolean;
-    showHintAndError: boolean;
-    showLabel: boolean;
-    title?: string;
-    value?: K;
+type injectedProps = 'searchValue' | 'onSearchChange' | 'searchOptions';
+type Props<T, K extends OptionKey> = Omit<RawSelectInputProps<T, K>, injectedProps> & {
     maxDisplayOptions?: number;
+};
+
+function filterAndSearch<T, K>(
+    options: T[],
+    labelSelector: (datum: T) => string | number,
+    searchValue: string,
+) {
+    return options
+        .filter(option => caseInsensitiveSubmatch(labelSelector(option), searchValue))
+        .sort((a, b) => compareStringSearch(
+            String(labelSelector(a)),
+            String(labelSelector(b)),
+            String(searchValue),
+        ));
 }
+
 function SelectInput<T = DefaultItem, K extends OptionKey = string>(props: Props<T, K>) {
     const {
-        autoFocus,
-        className: classNameFromProps,
-        disabled,
-        error,
-        hint,
-        keySelector,
-        label,
-        labelSelector,
         maxDisplayOptions,
-        onChange,
-        optionLabelSelector,
+        labelSelector,
         options,
-        optionsClassName,
-        placeholder = maxDisplayOptions === undefined
-            ? 'Select an option'
-            : 'Search for an option',
-        readOnly,
-        renderEmpty,
-        showClearButton: showClearButtonFromProps,
-        showDropdownArrowButton,
-        showHintAndError,
-        showLabel,
-        title,
-        value,
+        placeholder,
+        ...otherProps
     } = props;
 
-    const containerRef = useRef<HTMLDivElement>(null);
-    const inputRef = useRef<HTMLInputElement>(null);
-    const boundingClientRect = useRef<ClientRect>();
-
-    const [inputInFocus, setInputInFocus] = useState(false);
-    const [focusedKey, setFocusKey] = useState();
-    const [showOptionsPopup, setShowOptionsPopup] = useState(false);
     const [searchValue, setSearchValue] = useState<string | undefined>(undefined);
 
-    useEffect(
+    const searchOptions = useMemo(
         () => {
-            const { current: container } = containerRef;
-            if (container) {
-                boundingClientRect.current = container.getBoundingClientRect();
-            }
-        },
-        [containerRef.current],
-    );
-
-    const handleShowOptionsPopup = useCallback(
-        () => {
-            const { current: input } = inputRef;
-            if (input) {
-                input.select();
+            if (maxDisplayOptions === undefined) {
+                if (isFalsyString(searchValue)) {
+                    return options;
+                }
+                return filterAndSearch(options, labelSelector, searchValue);
             }
 
-            if (!showOptionsPopup) {
-                setShowOptionsPopup(true);
-                setSearchValue(undefined);
-            }
-        },
-        [inputRef, value, showOptionsPopup],
-    );
-
-    const handleToggleOptionsPopup = useCallback(
-        () => {
-            const { current: input } = inputRef;
-            if (input) {
-                input.select();
-            }
-
-            if (!showOptionsPopup) {
-                setShowOptionsPopup(true);
-                setSearchValue(undefined);
-            } else {
-                setShowOptionsPopup(false);
-            }
-        },
-        [inputRef, value, showOptionsPopup],
-    );
-
-    const handleHideOptionsPopup = useCallback(
-        () => {
-            setShowOptionsPopup(false);
-        },
-        [],
-    );
-
-    const handleInputFocus = useCallback(
-        () => {
-            setInputInFocus(true);
-        },
-        [],
-    );
-
-    const handleInputBlur = useCallback(
-        () => {
-            setInputInFocus(false);
-        },
-        [],
-    );
-
-    const handleInputChange = useCallback(
-        (e: React.FormEvent<HTMLInputElement>) => {
-            const { value: val } = e.currentTarget;
-            setShowOptionsPopup(true);
-            setSearchValue(val);
-        },
-        [],
-    );
-
-    const handleOptionsInvalidate = useCallback(
-        (optionsContainer: HTMLDivElement) => {
-            const contentRect = optionsContainer.getBoundingClientRect();
-
-            const { current: container } = containerRef;
-            const parentRect = container
-                ? container.getBoundingClientRect()
-                : boundingClientRect.current;
-
-            const offset = { ...defaultOffset };
-
-            const limit = { ...defaultLimit };
-            if (parentRect) {
-                limit.minW = parentRect.width;
-                limit.maxW = parentRect.width;
-            }
-
-            const optionsContainerPosition = calcFloatPositionInMainWindow({
-                parentRect,
-                contentRect,
-                offset,
-                limit,
-            });
-
-            return optionsContainerPosition;
-        },
-        [containerRef, boundingClientRect.current, showHintAndError],
-    );
-
-    const handleOptionSelect = useCallback(
-        (optionKey: K) => {
-            setShowOptionsPopup(false);
-            if (optionKey !== value) {
-                onChange(optionKey);
-            }
-        },
-        [value, onChange],
-    );
-
-    const handleClearButtonClick = useCallback(
-        () => {
-            if (value !== undefined) {
-                onChange(undefined);
-            }
-        },
-        [value, onChange],
-    );
-
-    const handleFocusChange = useCallback(
-        (key: K | undefined) => {
-            setFocusKey(key);
-        },
-        [],
-    );
-
-    const filteredOptions = useMemo(
-        () => {
             if (isFalsyString(searchValue)) {
-                return maxDisplayOptions !== undefined
-                    ? []
-                    : options;
+                return [];
             }
-            let newOptions = options
-                .filter(option => caseInsensitiveSubmatch(labelSelector(option), searchValue))
-                .sort((a, b) => compareStringSearch(
-                    String(labelSelector(a)),
-                    String(labelSelector(b)),
-                    String(searchValue),
-                ));
-            if (maxDisplayOptions !== undefined) {
-                newOptions = newOptions.slice(0, maxDisplayOptions);
-            }
-            return newOptions;
+
+            return filterAndSearch(options, labelSelector, searchValue)
+                .slice(0, maxDisplayOptions);
         },
         [options, labelSelector, maxDisplayOptions, searchValue],
     );
 
-    const defaultSearchValue = useMemo(
-        () => {
-            const activeOption = options.find(
-                d => keySelector(d) === value,
-            );
-
-            return activeOption !== undefined
-                ? String(labelSelector(activeOption))
-                : '';
-        },
-        [options, labelSelector, keySelector, value],
-    );
-
-    const finalSearchValue: string | undefined = !showOptionsPopup
-        ? defaultSearchValue
-        : searchValue;
-
-    const isFilled = isTruthyString(finalSearchValue);
-
-    const showClearButton = showClearButtonFromProps
-        && !disabled
-        && !readOnly
-        && isFilled;
-
-    const handleKeyDown = useKeyboard(
-        focusedKey,
-        value,
-        keySelector,
-        filteredOptions,
-        showOptionsPopup,
-
-        handleFocusChange,
-        handleHideOptionsPopup,
-        handleShowOptionsPopup,
-        handleOptionSelect,
-    );
-
-    const className = _cs(
-        classNameFromProps,
-        'select-input',
-        styles.selectInput,
-        showOptionsPopup && styles.showOptions,
-        showOptionsPopup && 'show-options',
-        disabled && styles.disabled,
-        disabled && 'disabled',
-        error && styles.error,
-        error && 'error',
-        inputInFocus && styles.inputInFocus,
-        inputInFocus && 'input-in-focus',
-        !showClearButton && styles.hideClearButton,
-        !showClearButton && 'hide-clear-button',
-        isFilled && styles.filled,
-        isFilled && 'filled',
-    );
+    const defaultPlaceholder = maxDisplayOptions === undefined
+        ? 'Select an option'
+        : 'Search for an option';
 
     return (
-        <div
-            className={className}
-            title={title}
-        >
-            { showLabel && (
-                <Label
-                    text={label}
-                    error={!!error}
-                    disabled={disabled}
-                    active={inputInFocus || showOptionsPopup}
-                />
-            )}
-            <div
-                ref={containerRef}
-                className={_cs(styles.inputAndActions, 'input-and-actions')}
-            >
-                <RawInput
-                    ref={inputRef}
-                    className={styles.input}
-                    type="text"
-                    onBlur={handleInputBlur}
-                    onFocus={handleInputFocus}
-                    onClick={handleShowOptionsPopup}
-                    onChange={handleInputChange}
-                    value={finalSearchValue || ''}
-                    autoFocus={autoFocus}
-                    placeholder={placeholder}
-                    disabled={disabled || readOnly}
-
-                    onKeyDown={handleKeyDown}
-                />
-                <div className={_cs('actions', styles.actions)}>
-                    { showClearButton && (
-                        <Button
-                            buttonType="button-danger"
-                            tabIndex={-1}
-                            iconName="close"
-                            className={_cs('clear-button', styles.clearButton)}
-                            onClick={handleClearButtonClick}
-                            transparent
-                        />
-                    )}
-                    { showDropdownArrowButton && (
-                        <Button
-                            tabIndex={-1}
-                            iconName={
-                                showOptionsPopup ? 'arrowDropup' : 'arrowDropdown'
-                            }
-                            className={_cs('dropdown-button', styles.dropdownButton)}
-                            onClick={handleToggleOptionsPopup}
-                            disabled={disabled || readOnly}
-                            transparent
-                        />
-                    )}
-                </div>
-            </div>
-            { showHintAndError && (
-                <HintAndError
-                    error={error}
-                    hint={hint}
-                />
-            )}
-            { showOptionsPopup && (
-                <Options
-                    className={optionsClassName}
-                    labelSelector={labelSelector}
-                    onBlur={handleHideOptionsPopup}
-                    onInvalidate={handleOptionsInvalidate}
-                    onOptionClick={handleOptionSelect}
-                    optionLabelSelector={optionLabelSelector}
-                    parentRef={containerRef}
-                    renderEmpty={renderEmpty}
-                    value={value}
-
-                    keySelector={keySelector}
-                    data={filteredOptions}
-                    onOptionFocus={handleFocusChange}
-                    focusedKey={focusedKey}
-                />
-            )}
-        </div>
+        <RawSelectInput
+            searchValue={searchValue}
+            onSearchChange={setSearchValue}
+            placeholder={placeholder || defaultPlaceholder}
+            searchOptions={searchOptions}
+            options={options}
+            labelSelector={labelSelector}
+            {...otherProps}
+        />
     );
 }
 SelectInput.defaultProps = {
-    disabled: false,
     keySelector: (item: DefaultItem) => item.key,
     labelSelector: (item: DefaultItem) => item.label,
     options: [],
-    readOnly: false,
-    showClearButton: true,
-    showDropdownArrowButton: true,
-    showHintAndError: true,
-    showLabel: true,
 };
 
 export default SelectInput;
